small changes to download data

+++ src/data/download_data.py
@@ -2,7 +2,6 @@ import os
 from typing import List
 from yahoofinancials import YahooFinancials
 import datetime
-import pandas as pd
 import csv
 
 destination_folder = os.path.abspath('./src/data/source')
@@ -10,12 +9,17 @@ csv_columns = ['date', 'high', 'low', 'open',
                'close', 'volume', 'adjclose', 'formatted_date']
 
 
-def download(codes: List[str]):
+def download(codes: List[str], allFlag: bool = False):
+    print("If data takes more than 10 seconds to download, to a ctrl + c to go to next stock")
     for code in codes:
         destination = os.path.join(destination_folder, code + '.csv')
-        if os.path.exists(destination):
-            response = input("would you like to overwrite that file? (y/n)")
-            if response == 'y' and response == 'yes':
+        if allFlag and os.path.exists(destination):
+            os.remove(destination)
+
+        if not allFlag and os.path.exists(destination):
+            response = input(
+                "would you like to overwrite that file(" + destination + ")? (y/n)")
+            if response == 'y' or response == 'yes':
                 os.remove(destination)
             else:
                 continue
@@ -24,11 +28,17 @@ def download(codes: List[str]):
 
         currentDate = datetime.date.today()
         initialDate = currentDate - datetime.timedelta(days=365.24*5)
+        data = []
+        try:
+            data = currentStock.get_historical_price_data(start_date=str(initialDate),
+                                                          end_date=str(
+                currentDate),
+                time_interval='daily')
+        except:
+            print("\nCould not download data for " +
+                  code + " (it probably doesn't exist")
+            continue
 
-        data = currentStock.get_historical_price_data(start_date=str(initialDate),
-                                                      end_date=str(
-                                                          currentDate),
-                                                      time_interval='daily')
         prices = data[code]['prices']
 
         with open(destination, 'w') as csvfile:
@@ -39,4 +49,7 @@ def download(codes: List[str]):
 
 
 if __name__ == '__main__':
-    download(['AAPL'])
+    array = ['^GSPTSE', '^N225', '^TNX', '^VIX', 'AAPL',
+             'ARL', 'BTC-USD', 'CL=F', 'GC=F', 'YVR']
+    download(array)
+    print("All data downloaded!")
